open_ai_practice/pyproject.toml
[tool.poetry]
name = "open-ai-practice"
version = "0.1.0"
description = ""
authors = ["Ananya Rana <rana.ananya.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
openai = "^1.55.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
